refactor(tournament.cache.ts): 토너먼트 캐시 파일 위치 변경 및 정리
